+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `RSpec/Be` cop to enforce passing argument to the generic `be` matcher. ([@Darhazer][])
 * Fix false positives in `StaticAttributeDefinedDynamically` and `ReturnFromStub` when a const is used in an array or hash. ([@Darhazer][])
 * Add `RSpec/Pending` cop to enforce no existing pending or skipped examples.  This is disabled by default. ([@patrickomatic][])
+* Fix `RSpec/NestedGroups` cop support --auto-gen-config. ([@walf443][])
 
 ## 1.24.0 (2018-03-06)